pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "psycopg2-binary>=2.9.10",
]
